+++ andersen-homework/src/main/java/org/example/dao/UserDao.java
@@ -1,11 +1,12 @@
 package org.example.dao;
 
-import org.example.model.User;
+
+import org.example.entity.User;
 
 public interface UserDao {
     void saveUser(User user);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     void deleteUserAndTicketsById(int userId);
 }
